+++ packages/config/src/configProcessor/helpers/azion.config.example.ts
@@ -1,21 +1,21 @@
 import type {
-  AzionConfig,
-  AzionConnector,
-  CacheByCookie,
-  CacheByQueryString,
-  ConnectorDnsResolution,
-  ConnectorHttpVersionPolicy,
-  ConnectorTransportPolicy,
-  ConnectorType,
-  CustomPageErrorCode,
-  CustomPageType,
-  NetworkListType,
-  WafSensitivity,
-  WafThreatType,
-  WorkloadInfrastructure,
-  WorkloadMTLSVerification,
-  WorkloadTLSCipher,
-  WorkloadTLSVersion,
+    AzionConfig,
+    AzionConnector,
+    CacheByCookie,
+    CacheByQueryString,
+    ConnectorDnsResolution,
+    ConnectorHttpVersionPolicy,
+    ConnectorTransportPolicy,
+    ConnectorType,
+    CustomPageErrorCode,
+    CustomPageType,
+    NetworkListType,
+    WafSensitivity,
+    WafThreatType,
+    WorkloadInfrastructure,
+    WorkloadMTLSVerification,
+    WorkloadTLSCipher,
+    WorkloadTLSVersion,
 } from 'azion/config';
 
 const config: AzionConfig = {
@@ -411,7 +411,7 @@ const config: AzionConfig = {
   storage: [
     {
       name: 'my-storage',
-      edgeAccess: 'read_write', // 'read_only' | 'read_write' | 'restricted'
+      workloadsAccess: 'read_write', // 'read_only' | 'read_write' | 'restricted'
       dir: './storage',
       prefix: '1236677364374',
     },